+++ src/runtime/GrapheneApp.h
@@ -1,10 +1,6 @@
 #if !defined(GRAPHENE_APP_INTERFACE_H)
 #define GRAPHENE_APP_INTERFACE_H
 
-#include <map>
-
-#include "../core/Font.h"
-#include "../core/String.h"
 #include "../interfaces/AbstractCanvas.h"
 #include "../interfaces/AbstractTouchInput.h"
 #include "../interfaces/IContainer.h"
@@ -19,18 +15,6 @@ class GrapheneApp {
 
 	std::vector<TouchEvent> processTouchEvents(std::vector<Point> points);
 
-	std::vector<Graphene::String> getSystemFonts() {
-		std::vector<Graphene::String> fontNames = {};
-		for (auto &pair : GrapheneSystemFonts) {
-			fontNames.push_back(pair.first);
-		}
-		return fontNames;
-	}
-
-	Graphene::Font &getSystemFont(Graphene::String fontName) {
-		return GrapheneSystemFonts[fontName];
-	}
-
    protected:
 	AbstractCanvas *canvas;
 	AbstractTouchInput *touchInput;
@@ -38,9 +22,6 @@ class GrapheneApp {
 
 	bool hasPrevTouchEvent = false;
 	Graphene::TouchEvent prevTouchEvent;
-
-   private:
-	std::map<Graphene::String, Graphene::Font> GrapheneSystemFonts;
 };
 
 }  // namespace Graphene